The programme works like a charm ;]
It's a continuation of making changes in the code to make it more clear and less repetitive.
The Programme can connect to database and do some CRUD operation like adding, getting objects by ID, updating info and deleting. Functions concerns both books and clients. Additionally you can rent and return a book, check its availability and a date of return if it's rented.